More work
Add a schematic for the decoupling and the passive components for the standalone mode (simple switches)
The lock_handler.sch schematic is in RC state.
PCB: Daisy chain passive components completed.
PCB: Estimation of the passive components of the switches.
PCB: complete Vcc and ground planes, including some high frequency printed low value decoupling.
The top of the PCB is in RC state.